# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Your Overtime API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/holiday": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holiday"
                ],
                "summary": "Get a activities by start and end",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.Holiday"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holiday"
                ],
                "summary": "Creates a holiday",
                "parameters": [
                    {
                        "description": "input holiday",
                        "name": "bottles",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.InputHoliday"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Holiday"
                        }
                    }
                }
            }
        },
        "/holiday/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holiday"
                ],
                "summary": "Get a holiday by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Holiday id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Holiday"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holiday"
                ],
                "summary": "Updates a holiday",
                "parameters": [
                    {
                        "description": "input holiday",
                        "name": "bottles",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.InputHoliday"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Holiday id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Holiday"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holiday"
                ],
                "summary": "Delete a holiday",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Holiday id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Holiday"
                        }
                    }
                }
            }
        },
...
        "/token": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Retrieves tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.Token"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "creates a token",
                "parameters": [
                    {
                        "description": "input token",
                        "name": "bottles",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.InputToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Token"
                        }
                    }
                }
            }
        },
        "/token/{id}": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Delete a token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Token"
                        }
                    }
                }
            }
        },
        "/workday": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workday"
                ],
                "summary": "Retrieves workdays",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.WorkDay"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workday"
                ],
                "summary": "creates a workdays",
                "parameters": [
                    {
                        "description": "input workday",
                        "name": "bottles",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.InputWorkDay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.WorkDay"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "pkg.Holiday": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "start": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "pkg.InputHoliday": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "pkg.InputToken": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "pkg.InputWorkDay": {
            "type": "object",
            "properties": {
                "activeTime": {
                    "type": "integer"
                },
                "day": {
                    "type": "string"
                },
                "overtime": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "pkg.Token": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "pkg.WorkDay": {
            "type": "object",
            "properties": {
                "activeTime": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isHoliday": {
                    "type": "boolean"
                },
                "overtime": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "type": "apiKey",
            "name": "adminToken",
            "in": "query"
        },
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}